requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==1.11.17
django-crispy-forms==1.13.0
djangorestframework==3.11.2
gunicorn==20.1.0
jsonfield==3.1.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
